requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
brotlipy=0.7.0=py38h27cfd23_1003
ca-certificates=2020.10.14=0
certifi=2020.6.20=pyhd3eb1b0_3
cffi=1.14.3=py38h261ae71_2
chardet=3.0.4=py38h06a4308_1003
conda=4.9.2=py38h06a4308_0
conda-package-handling=1.7.2=py38h03888b9_0
cryptography=3.2.1=py38h3c74f83_1
funcsigs=1.0.2=pypi_0
idna=2.10=py_0
imageio=2.9.0=pypi_0
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
mypy-extensions=0.4.3=pypi_0
ncurses=6.2=he6710b0_1
numpy=1.19.4=pypi_0
openssl=1.1.1h=h7b6447c_0
packaging=20.8=pypi_0
pandas=1.1.5=pypi_0
pillow=8.0.1=pypi_0
pip=20.2.4=py38h06a4308_0
pycosat=0.6.3=py38h7b6447c_1
pycparser=2.20=py_2
pyopenssl=19.1.0=pyhd3eb1b0_1
pyparsing=2.4.7=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.5=h7579374_1
python-dateutil=2.8.1=pypi_0
pytz=2020.4=pypi_0
readline=8.0=h7b6447c_0
regex=2020.11.13=pypi_0
requests=2.24.0=py_0
ruamel_yaml=0.15.87=py38h7b6447c_1
sentencepiece=0.1.94=pypi_0
setuptools=50.3.1=py38h06a4308_1
six=1.15.0=py38h06a4308_0
sqlite=3.33.0=h62c20be_0
texar-pytorch=0.1.1=pypi_0
tk=8.6.10=hbc83047_0
torch=1.7.1=pypi_0
torchvision=0.8.2=pypi_0
tqdm=4.51.0=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.25.11=py_0
wheel=0.35.1=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
